+++ RELEASE-NOTES.md
@@ -6,6 +6,7 @@ API Details can always be found in the [wiki](https://github.com/JeffersonLab/E2
 - Improved the detection of source address for sender registration - `--autoip` for senders should work reliably now.
 - Added SegmenterFlags.eventQueueSize parameter (still defaults to 2047) which governs the size of the queue from which the events are sent out. In case of a host running out of RAM, reduce this.
 - Modified the default ReassemblerFlags.reportStats to be true (it was set to false for compatibility reasons with older CPs, however doesn't seem relevant anymore; in the future this flag will be deprecated)
+- Deprecated ReassemblerFlags.withLBHeader - the field is present, but should no longer be modified - it is always set to 'not rFlags.useCP', i.e. when control plane is not in use.
 
 ## v0.3.2
 - Fixed Conda packaging